examples/crewai/requirements.txt
crewai>=0.152.0
crewai-tools[mcp]>=0.39.0
python-dotenv>=1.0.0
